requirements.txt
cassandra-driver
python-dotenv
faker
fastapi
uvicorn
